requirements.txt
python-dotenv
Flask
Flask-Login
validators
steam
urllib3
requests
flask_wtf
wtforms
sqlalchemy
werkzeug
sqlalchemy_serializer